Rename openssl.ts cryptography.ts
Since we could one day use something other than OpenSSL for
cryptographic operations
