+++ libs/auth/scripts/read_java_card_details.ts
@@ -5,9 +5,9 @@ import {
   PROD_VX_CERT_AUTHORITY_CERT_PATH,
 } from '../src';
 import { CardDetails } from '../src/card';
+import { verifyFirstCertWasSignedBySecondCert } from '../src/cryptography';
 import { JavaCard } from '../src/java_card';
 import { waitForReadyCardStatus } from './utils';
-import { verifyFirstCertWasSignedBySecondCert } from '../src/openssl';
 
 const ENVS = ['development', 'production'] as const;
